12: Use AuthContext across the app
The app now has the AuthContextProvider to provide the "state" of the
user.

- If user is not logged in or signed up then 2 options are shown in the
center of the screen to sugn up/log in with a social account with Facebook
and Google being supported at the moment.

- While user is not authenticated the user will keep being redirected to
sign up.

- As long as the user is authenticated then a "Signout" button appears
in the "Navbar" and favorites appears. These 2 components of course
dissappear when user is not logged in.